projects/image_captioner/requirements.txt
torch>=2.0.0
transformers>=4.30.0
pillow>=10.0.0
flask>=3.0.0

